backend/requirements/requirements-base.txt
# django, rest-framework
# =====================
Django>=3.2.13,<4
django-filter==21.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
drf-spectacular==0.22.1
Markdown==3.3.7
python-dotenv==0.20.0
drf-nested-routers==0.93.4
django-model-utils==4.2.0

# DI
# =====================
dependency-injector==4.39.1
